declare dtypes for pandas Dataframes upon creation creating directories before multiprocessing to avoid confusion among workers updated version

+++ temoatools/analyze_activity_year.py
@@ -44,7 +44,7 @@ def getActivity(folders, dbs, switch='fuel', sector_name='electric', save_data='
         folders = fldrs
 
     # Create dictionary to hold each capacity_single series
-    activity = pd.DataFrame()
+    activity = pd.DataFrame(dtype='float64')
 
     # Iterate through each db
     for folder, db in zip(folders, dbs):
@@ -171,9 +171,6 @@ def SingleDB(folder, db, switch='fuel', sector_name='electric', conversion=277.7
     future_t_periods = sorted(future_t_periods)
     rows = future_t_periods[:-1]  # Last period is not calculated
 
-    # Create dataframe initialized to zero
-    # df = pd.DataFrame(data=0.0,index=rows,columns = cols)
-
     #   Identify Unique Scenarios
     qry = "SELECT * FROM Output_Objective"
     cur.execute(qry)
@@ -185,7 +182,7 @@ def SingleDB(folder, db, switch='fuel', sector_name='electric', conversion=277.7
 
     # Create pandas DataFrame to hold yearlyEmissions for all scenarios
     index = pd.MultiIndex.from_product([[db], scenarios, cols], names=['database', 'scenario', 'fuelOrTech'])
-    df = pd.DataFrame(index=index, columns=future_t_periods[:-1])
+    df = pd.DataFrame(index=index, columns=future_t_periods[:-1], dtype='float64')
     df = df.fillna(0.0)  # Default value to zero
 
     ## Review db_Output_VFlow_Out to fill data frame
